refactor: deduplicate aggregation and date formatting utilities
Remove duplicated aggregateByKey() from usage-helpers.js — now imports
from aggregation-utils.js. Centralize formatTime/formatDateTime in
shared/date-utils.js, with main/date-utils.js and src/utils/date-utils.js
both delegating to the single implementation.

Closes #59
