+++ main/date-utils.js
@@ -1,10 +1,15 @@
 /**
- * Shared date/time utilities used across main and renderer modules.
+ * Date/time utilities for the main process.
  * Pure functions — no side effects.
+ *
+ * Generic formatting (formatTime, formatDateTime) lives in shared/date-utils.js
+ * and is re-exported here for backward compatibility.
+ * Domain-specific helpers (extractDateString, generateDateRange) stay here.
  */
 
+const { formatTime, formatDateTime } = require('../shared/date-utils');
+
 const DATE_LOCALE = 'fr-FR';
-const TIME_FORMAT = { hour: '2-digit', minute: '2-digit' };
 const DAY_LABEL_FORMAT = { day: '2-digit', month: '2-digit' };
 
 /**
@@ -31,30 +36,6 @@ function generateDateRange(days = 30) {
       label: d.toLocaleDateString(DATE_LOCALE, DAY_LABEL_FORMAT),
     };
   });
-}
-
-/**
- * Format a timestamp into a short time string (e.g. "14:32").
- * Returns '' if timestamp is falsy.
- * @param {number|string|null} timestamp
- * @returns {string}
- */
-function formatTime(timestamp) {
-  return timestamp
-    ? new Date(timestamp).toLocaleTimeString(DATE_LOCALE, TIME_FORMAT)
-    : '';
-}
-
-/**
- * Build a "date time" label from a date string and optional timestamp.
- * e.g. "2025-03-29 14:32" or just "2025-03-29" if no timestamp.
- * @param {string} date - date string (e.g. "2025-03-29")
- * @param {number|string|null} timestamp
- * @returns {string}
- */
-function formatDateTime(date, timestamp) {
-  const time = formatTime(timestamp);
-  return `${date}${time ? ' ' + time : ''}`;
 }
 
 module.exports = { extractDateString, generateDateRange, formatTime, formatDateTime };
